+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppssubjectaccessrequestworker/services/GeneratePdfServiceTest.kt
@@ -1,7 +1,13 @@
 package uk.gov.justice.digital.hmpps.hmppssubjectaccessrequestworker.services
 
+import com.itextpdf.text.BaseColor
+import com.itextpdf.text.Chunk
 import com.itextpdf.text.Document
+import com.itextpdf.text.FontFactory
+import com.itextpdf.text.pdf.ColumnText
+import com.itextpdf.text.pdf.PdfReader
 import com.itextpdf.text.pdf.PdfWriter
+import com.itextpdf.text.pdf.parser.PdfTextExtractor
 import io.kotest.core.spec.style.DescribeSpec
 import org.assertj.core.api.Assertions
 import org.mockito.Mockito
@@ -13,6 +19,7 @@ import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.context.ContextConfiguration
 import uk.gov.justice.digital.hmpps.hmppssubjectaccessrequestworker.models.CustomHeader
 import java.io.ByteArrayOutputStream
+import java.io.FileOutputStream
 
 @ActiveProfiles("test")
 @ContextConfiguration(
@@ -42,7 +49,7 @@ class GeneratePdfServiceTest(
         generatePdfService.execute(testResponseObject, "", "", mockDocument, mockStream)
 
         verify(mockDocument, Mockito.times(1)).open()
-        verify(mockDocument, Mockito.times(1)).add(any())
+        verify(mockDocument, Mockito.times(3)).add(any())
         verify(mockDocument, Mockito.times(1)).close()
       }
 
@@ -54,65 +61,34 @@ class GeneratePdfServiceTest(
         val stream = generatePdfService.execute(testResponseObject, "", "", mockDocument, mockStream)
         Assertions.assertThat(stream).isInstanceOf(ByteArrayOutputStream::class.java)
       }
-      it("returns a ByteArrayOutputStream") {
-        val testResponseObject: Map<String, Any> = mapOf("Dummy" to "content")
-        val mockDocument = Mockito.mock(Document::class.java)
-        val mockPdfService = Mockito.mock(PdfService::class.java)
-        val mockStream = Mockito.mock(ByteArrayOutputStream::class.java)
-        val mockWriter = Mockito.mock(PdfWriter::class.java)
-        Mockito.`when`(mockPdfService.getPdfWriter(mockDocument, mockStream)).thenReturn(mockWriter)
-        Mockito.`when`(mockWriter.pageNumber).thenReturn(1)
-        val stream = generatePdfService.execute(testResponseObject, "NDELIUS ID: EGnDeliusID", "EGsarID", mockDocument, mockStream)
-        Assertions.assertThat(stream).isInstanceOf(ByteArrayOutputStream::class.java)
-      }
 
-      it("calls iText open, add and close") {
-        val testResponseObject: Map<String, Any> =
-          mapOf("content" to mapOf<String, Any>("fake-prisoner-search-property" to emptyMap<String, Any>()))
-        val mockDocument = Mockito.mock(Document::class.java)
-        val mockPdfService = Mockito.mock(PdfService::class.java)
-        val mockPdfWriter = Mockito.mock(PdfWriter::class.java)
-        val mockStream = Mockito.mock(ByteArrayOutputStream::class.java)
-        val mockHeader = Mockito.mock(CustomHeader::class.java)
-        Mockito.`when`(mockPdfService.getPdfWriter(mockDocument, mockStream)).thenReturn(mockPdfWriter)
-        Mockito.`when`(mockPdfService.getCustomHeader("NDELIUS ID: EGnDeliusID", "EGsarID")).thenReturn(mockHeader)
-        Mockito.`when`(mockPdfService.setEvent(mockPdfWriter, mockHeader)).thenReturn(0)
-
-        generatePdfService.execute(testResponseObject, "NDELIUS ID: EGnDeliusID", "EGsarID", mockDocument, mockStream)
-        verify(mockPdfService, Mockito.times(1)).getPdfWriter(mockDocument, mockStream)
-        verify(mockDocument, Mockito.times(1)).open()
-        verify(mockDocument, Mockito.times(1)).add(any())
-        verify(mockDocument, Mockito.times(1)).close()
-      }
       it("adds header to a PDF") {
         val testResponseObject: Map<String, Any> =
           mapOf("content" to mapOf<String, Any>("fake-prisoner-search-property" to emptyMap<String, Any>()))
         val mockDocument = Mockito.mock(Document::class.java)
-        val mockPdfService = Mockito.mock(PdfService::class.java)
-        val mockPdfWriter = Mockito.mock(PdfWriter::class.java)
+        val mockCustomHeader = Mockito.mock(CustomHeader::class.java)
         val mockStream = Mockito.mock(ByteArrayOutputStream::class.java)
         val mockWriter = Mockito.mock(PdfWriter::class.java)
-        val mockHeader = Mockito.mock(CustomHeader::class.java)
-        Mockito.`when`(mockPdfService.getCustomHeader("NDELIUS ID: EGnDeliusID", "EGsarID")).thenReturn(mockHeader)
-        Mockito.`when`(mockPdfService.setEvent(mockPdfWriter, mockHeader)).thenReturn(0)
-        Mockito.`when`(mockPdfService.getPdfWriter(mockDocument, mockStream)).thenReturn(mockWriter)
-        Mockito.`when`(mockWriter.pageNumber).thenReturn(1)
-        generatePdfService.execute(testResponseObject, "NDELIUS ID: EGnDeliusID", "EGsarID", mockDocument, mockStream)
-        verify(mockPdfService, Mockito.times(1)).getCustomHeader("NDELIUS ID: EGnDeliusID", "EGsarID")
+        val mockTextAligner = Mockito.mock(ColumnText::class.java)
+        val stream = generatePdfService.execute(testResponseObject, "NDELIUS ID: EGnDeliusID", "EGsarID", mockDocument, mockStream)
+        Assertions.assertThat(stream.toByteArray()).contains("EGnDeliusID".toByteArray())
       }
 
-      it("adds rear page") {
-        val testResponseObject = mutableMapOf<String, Any>()
-        val mockDocument = Mockito.mock(Document::class.java)
-        val mockPdfService = Mockito.mock(PdfService::class.java)
-        val mockStream = Mockito.mock(ByteArrayOutputStream::class.java)
-        val mockWriter = Mockito.mock(PdfWriter::class.java)
-        Mockito.`when`(mockPdfService.getPdfWriter(mockDocument, mockStream)).thenReturn(mockWriter)
-        Mockito.`when`(mockWriter.pageNumber).thenReturn(1)
-        generatePdfService.execute(testResponseObject, "NDELIUS ID: EGnDeliusID", "EGsarID", mockDocument, mockStream)
-        verify(mockPdfService, Mockito.times(1)).addRearPage(document = any(), font = any(), numPages = any())
+      it("adds rear page with correct text") {
+        val mockDocument = Document()
+        val writer = PdfWriter.getInstance(mockDocument, FileOutputStream("dummy.pdf"))
+        mockDocument.open()
+        val font = FontFactory.getFont(FontFactory.COURIER, 20f, BaseColor.BLACK)
+        mockDocument.add(Chunk("Text so that the page isn't empty", font))
+        writer.isPageEmpty = false
+        Assertions.assertThat(writer.pageNumber).isEqualTo(1)
+        generatePdfService.addRearPage(mockDocument, writer.pageNumber)
+        mockDocument.close()
+        val reader = PdfReader("dummy.pdf")
+        val text = PdfTextExtractor.getTextFromPage(reader, 2)
+        Assertions.assertThat(text).contains("End of Subject Access Request Report")
+        Assertions.assertThat(text).contains("Total pages: 1")
       }
-
     }
   },
 )
